chore: re-add equivalent of old `ForeignCallResult::From` implementation (#391)
